requirements.txt
uvicorn==0.34.3
PyYAML==6.0.2
gdown==5.2.0
langchain>=0.3.26
langchain-community>=0.3.27
langchain-huggingface>=0.3.0
langchain-chroma>=0.2.4
sentence-transformers>=4.1.0
redis>=5.2.1
chromadb>=1.0.13
python-dotenv>=1.0.0
openai>=0.27.8
langgraph>=0.1.2
celery[redis]>=5.5.3
openai>=0.27.8
python-dotenv>=1.0.0
requests==2.32.4
tqdm==4.67.1
sentence-transformers==4.1.0
python-docx==1.2.0
